requirements.txt
wxpython==4.1.1
pyyaml==5.4.1